requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
idna==3.1
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
matplotlib==3.4.2
numba==0.53.1
numpy==1.20.3
packaging==21.0
pandas==1.2.5
pillow==8.3.1
pip==21.1.3
pooch==1.4.0
pycparser==2.20
pydub==0.25.1
pyopenssl==20.0.1
pyparsing==2.4.7
pysocks==1.7.1
pysoundfile==0.10.3.post1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.6.2
setuptools==52.0.0
six==1.16.0
threadpoolctl==2.2.0
tornado==6.1
typing_extensions==3.10.0.0
urllib3==1.26.6
wheel==0.36.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
